Consolidated conditional branch logic, fixed cycle counts
There is one issue. Suppose that a branch needs 4 cycles to complete,
and that the branch is to be taken. Currently, the emulator evaluates
the branch at the 4th cycle and delays the following instruction by one
(or two) cycle(s). This is techincally incorrect.

Ultimately, this shouldn't exhibit any errors in the graphical
display, since simply moving the PC will not affect anything on the
display, and the next instruction (which could affect the display) is
still executed at the right time.